refactor: move test files into new folders to match src
